[MODULES] Add radio encryption and isolate temperature polling

+++ temperature-sensor-example/src/temperature_sensor_example_internal.h
@@ -10,7 +10,7 @@
  * The cache and the sensor are separate files so the cache can be tested on a
  * board that has no die temperature, which is every board Twister runs on.
  */
-void kfsw_temp_example_store(int32_t milli_c, uint32_t monotonic_ms);
+void kfsw_temp_example_store(int32_t milli_c, uint64_t monotonic_ms);
 void kfsw_temp_example_store_failure(void);
 
 #if CONFIG_KFSW_TEMP_EXAMPLE_SENSOR
